# gest-ashoex-backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Health Check",
        "description": "API para verificar el estado de salud del sistema",
        "version": "1.0.0"
    },
    "paths": {
        "/api/aulas": {
            "get": {
                "tags": [
                    "Aulas"
                ],
                "summary": "Obtener todas las aulas con sus usos y facilidades",
                "description": "Retorna una lista de todas las aulas junto con sus usos y facilidades.",
                "operationId": "getAllAulas",
                "responses": {
                    "200": {
                        "description": "Éxito. Lista de aulas obtenida correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "numero_aula": {
                                                "type": "string",
                                                "example": "A101"
                                            },
                                            "capacidad": {
                                                "type": "integer",
                                                "example": 40
                                            },
                                            "habilitada": {
                                                "type": "boolean",
                                                "example": true
                                            },
                                            "id_ubicacion": {
                                                "type": "integer",
                                                "example": 2
                                            },
                                            "usos": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "nombre": {
                                                            "type": "string",
                                                            "example": "Clases"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "facilidades": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "integer",
                                                            "example": 1
                                                        },
                                                        "descripcion": {
                                                            "type": "string",
                                                            "example": "Proyector"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Aulas"
                ],
                "summary": "Almacena una nueva aula en la base de datos",
                "operationId": "4ed9f79dd5cd9461861129d4b77d29db",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero_aula",
                                    "id_ubicacion",
                                    "id_uso",
                                    "facilidades"
                                ],
                                "properties": {
                                    "numero_aula": {
                                        "type": "string",
                                        "maxLength": 30,
                                        "example": "Aula 101"
                                    },
                                    "capacidad": {
                                        "type": "integer",
                                        "example": 30,
                                        "nullable": true
                                    },
                                    "habilitada": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "id_ubicacion": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "id_uso": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "facilidades": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Aula registrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "numero_aula": {
                                                    "type": "string",
                                                    "example": "Aula 101"
                                                },
                                                "capacidad": {
                                                    "type": "integer",
                                                    "example": 30
                                                },
                                                "habilitada": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "id_ubicacion": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "id_uso": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "facilidades": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "error": {
                                            "type": "array",
                                            "items": {},
                                            "example": []
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Aula registrada exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {},
                                            "example": []
                                        },
                                        "error": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "status": {
                                                        "type": "string",
                                                        "example": "422"
                                                    },
                                                    "detail": {
                                                        "type": "string",
                                                        "example": "El campo numero aula es obligatorio."
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "status": "422",
                                                    "detail": "El campo numero aula es obligatorio."
                                                },
                                                {
                                                    "status": "422",
                                                    "detail": "El campo id ubicacion seleccionado no existe."
                                                },
                                                {
                                                    "status": "422",
                                                    "detail": "El campo facilidades seleccionado no existe."
                                                }
                                            ]
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}